\name{UM}
\docType{data}
\alias{UM}
\alias{mut02}
\alias{cn}
\alias{methy25}
\alias{mrna25}
\alias{methy25Anno}
\alias{clin4}
\title{Uveal Melanoma data}
\description{
 This is a subset of the uveal melanoma (UM) multi-omics data from the cancer
 genome atlas (TCGA) study (2018), which were re-analyzed by Mo et al. (2021).
}
\usage{data(UM)}
\format{
  Data matrices of somatic mutation, DNA copy number, methylation and
  mRNA expression for 80 UM primary samples.
} 
\value{
  \item{mut02}{Somatic mutation data matrix with 0 representing wild type and 1
  representing somatic mutation. Genes with mutation rate >= 2\% in the
  80 samples are kept in the data matrix.}
  \item{cn}{Copy number regions, which were generated by merging the log2 ratios of
    chromosome segments using the CNregions function.} 
  \item{methy25}{Methylation data matrix made of the top 25\% most
    variable genes.}
  \item{mrna25}{mRNA expression data matrix made of the top 25\% most
    variable genes.}
  \item{methy25Anno}{Annotation for the genes in the methylation data
    matrix methy25.}
  \item{clin4}{Clinical data of the 80 UM samples.}
}
\details{
  The TCGA UM multi-omics data (version 2016_01_28) were obtained from
  the Firebrowse portal (http://firebrowse.org/, accessed on 19 December
  2018).  The level 3 multi-omics data were processed for iCluster
  analysis, which were detailed in the Materials and Methods of Mo et
  al. (2021).   
}

\references{
Robertson, A.G.; Shih, J.; Yau, C.; Gibb, E.A.; Oba, J.; Mungall, K.L.;
Hess, J.M.; Uzunangelov, V.;Walter, V.; Danilova, L.; et al. Integrative
Analysis Identifies Four Molecular and Clinical Subsets in Uveal
Melanoma. 2018.  \emph{Cancer Cell} 33 (1), 151. 


Mo, Q; Wan, L; Schell, MJ; Jim, H; Tworoger, SS; G Peng, G.  Integrative
analysis identifies multi-omics signatures that drive molecular
classification of uveal melanoma. 2021. \emph{Cancers} 13 (24), 6168
}
\keyword{datasets}
